ref(asyncpg): Move query whitespace normalization into integration
Instead of normalizing span descriptions generically in client.py
before calling before_send_transaction, normalize queries at the source
in the asyncpg integration. This is more correct because the
normalization is specific to SQL query formatting, not a general
concern for all span descriptions.

Remove _clean_span_descriptions from _Client and the generic tests.
Add asyncpg-specific tests for normalized descriptions and
before_send_transaction interaction.
